services/gemini-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-agent"
version = "0.1.0"
description = "Vertex AI Agent Builder webhook — bridges Gemini Enterprise to Ancol MoM Compliance pipeline"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "google-auth>=2.28",
    "google-cloud-spanner>=3.40",
    "ancol-common",
]

[project.optional-dependencies]
neo4j = ["neo4j>=5.15"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21"]
